osdconfig node enumeration added
