+++ osdconfig/api.go
@@ -4,6 +4,8 @@ import (
 	"encoding/json"
 	"fmt"
 	"path/filepath"
+
+	"github.com/pkg/errors"
 )
 
 // GetClusterConf retrieves cluster level data from kvdb
@@ -24,16 +26,28 @@ func (manager *configManager) GetClusterConf() (*ClusterConfig, error) {
 
 // SetClusterConf sets cluster config in kvdb
 func (manager *configManager) SetClusterConf(config *ClusterConfig) error {
-	manager.Lock()
-	defer manager.Unlock()
-
 	if config == nil {
 		return fmt.Errorf("input cannot be nil")
 	}
 
+	manager.Lock()
+	defer manager.Unlock()
+
+	key, err := manager.cc.Lock(baseKey)
+	if err != nil {
+		return err
+	}
+
 	// push into kvdb
-	_, err := manager.cc.Put(filepath.Join(baseKey, clusterKey), config, 0)
-	return err
+	if _, err := manager.cc.Put(filepath.Join(baseKey, clusterKey), config, 0); err != nil {
+		return err
+	}
+
+	if err := manager.cc.Unlock(key); err != nil {
+		return err
+	}
+
+	return nil
 }
 
 // GetNodeConf retrieves node config data using nodeID
@@ -58,9 +72,6 @@ func (manager *configManager) GetNodeConf(nodeID string) (*NodeConfig, error) {
 
 // SetNodeConf sets node config data in kvdb
 func (manager *configManager) SetNodeConf(config *NodeConfig) error {
-	manager.Lock()
-	defer manager.Unlock()
-
 	if config == nil {
 		return fmt.Errorf("input cannot be nil")
 	}
@@ -69,23 +80,71 @@ func (manager *configManager) SetNodeConf(config *NodeConfig) error {
 		return fmt.Errorf("node id cannot be nil")
 	}
 
+	manager.Lock()
+	defer manager.Unlock()
+
+	key, err := manager.cc.Lock(baseKey)
+	if err != nil {
+		return err
+	}
+
 	// push node data into kvdb
-	_, err := manager.cc.Put(getNodeKeyFromNodeID(config.NodeId), config, 0)
-	return err
+	if _, err := manager.cc.Put(getNodeKeyFromNodeID(config.NodeId), config, 0); err != nil {
+		return err
+	}
+
+	if err := manager.cc.Unlock(key); err != nil {
+		return err
+	}
+
+	return nil
 }
 
 // UnsetNodeConf deletes node config data in kvdb
 func (manager *configManager) UnsetNodeConf(nodeID string) error {
-	manager.Lock()
-	defer manager.Unlock()
-
 	if len(nodeID) == 0 {
 		return fmt.Errorf("node id cannot be nil")
 	}
 
+	manager.Lock()
+	defer manager.Unlock()
+
+	key, err := manager.cc.Lock(baseKey)
+	if err != nil {
+		return err
+	}
+
 	// remove dode data from kvdb
-	_, err := manager.cc.Delete(getNodeKeyFromNodeID(nodeID))
-	return err
+	if _, err := manager.cc.Delete(getNodeKeyFromNodeID(nodeID)); err != nil {
+		return err
+	}
+
+	if err := manager.cc.Unlock(key); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// EnumerateConf fetches data for all nodes
+func (manager *configManager) EnumerateConf() (*NodesConfig, error) {
+	keys, err := manager.cc.Keys(baseKey, nodeKey)
+	if err != nil {
+		return nil, errors.New("kvdb.Keys() returned error: " + err.Error())
+	}
+
+	nodesConfig := new(NodesConfig)
+	*nodesConfig = make([]*NodeConfig, len(keys))
+	for i, key := range keys {
+		key := key
+		nodeConfig, err := manager.GetNodeConf(key)
+		if err != nil {
+			return nil, err
+		}
+		(*nodesConfig)[i] = nodeConfig
+	}
+
+	return nodesConfig, nil
 }
 
 // WatchCluster registers user defined function as callback and sets a watch for changes
